feature/ID-05/AdminPanelFormJSXCreation
