--- src/app/add-product/loading.jsx
@@ -1,5 +1,0 @@
-const loading = () => {
-    return <div>Loading ...</div>
-  }
-  
-  export default loading
